move depobs.scanner.docker to depobs.docker
