+++ tests/fixtures/mozilla_services_channelserver_79157df7b193857a2e7e3fe8e61e38305e1d47d4_cargo_metadata_output.out
